:sparkles: 기술스택 조회 기능 추가

+++ src/main/java/org/example/develociraptor/global/idclass/ProjectTechStackId.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @Getter
 public class ProjectTechStackId implements Serializable {
 
-	private Long techStackId;
+	private Short techStackId;
 
 	private Long feedId;
 }
